% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capValues}
\alias{capValues}
\title{Cap values}
\usage{
capValues(x, cap.at.quantile = 0.02)
}
\arguments{
\item{x}{a numerical vector}

\item{cap.at.quantile}{a quantile at which variables are capped
to remove outliers.}
}
\value{
a numerical vector containing capped values
}
\description{
This function takes a numerical vector as an input and returns
values that are capped at a specific quantile.
This is useful for removing outliers.
}
