% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{computeDistances}
\alias{computeDistances}
\title{Compute distances}
\usage{
computeDistances(x)
}
\arguments{
\item{x}{a numerical matrix or vector.}
}
\value{
a numerical matrix
}
\description{
Compute distances
}
