% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{deleteBranch}
\alias{deleteBranch}
\title{Get p-values representing the influence of deleting branches of metacells}
\usage{
deleteBranch(
  Yg,
  Xt,
  Yj,
  dend,
  alternative,
  metacell.celltype,
  seed = 1234,
  ...
)
}
\arguments{
\item{Yg}{a numeric vector containing RNA expression values.}

\item{Xt}{a numeric vector containing chromatin accessibility.}

\item{Yj}{a numeric vector containing TF expression.}

\item{dend}{a dendrogram class object.}

\item{alternative}{a character string specifying the alternative hypothesis.
This must be one of "two.sided", "greater", and "less".}

\item{metacell.celltype}{a character vector specifying cell types of the metacells.}

\item{seed}{an integer}

\item{...}{further arguments to be passed.}
}
\value{
A data frame containing p-values.
}
\description{
This function computes sampling-based p-values representing influences
of deleting branches of metacells on the linear interaction model.
It requires a dendrogram of metacells.
}
