% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{deleteCellType}
\alias{deleteCellType}
\title{Get p-values representing the influence of deleting a celltype}
\usage{
deleteCellType(Yg, Xt, Yj, metacell.celltype, alternative, seed = 1234, ...)
}
\arguments{
\item{Yg}{a numeric vector containing RNA expression values.}

\item{Xt}{a numeric vector containing chromatin accessibility.}

\item{Yj}{a numeric vector containing TF expression.}

\item{metacell.celltype}{a character vector specifying cell types of the metacells.}

\item{alternative}{a character string specifying the alternative hypothesis.
This must be one of "two.sided", "greater", and "less".}

\item{seed}{an integer}

\item{...}{further arguments to be passed.}
}
\value{
A data frame containing p-values.
}
\description{
This function computes sampling-based p-values representing the influence
of deleting metacells of a given cell type on the linear interaction model.
}
