% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{deleteNeighbor}
\alias{deleteNeighbor}
\title{Get p-values representing the influence of deleting a group of metacells}
\usage{
deleteNeighbor(
  Yg,
  Xt,
  Yj,
  alternative,
  metacell.celltype,
  metacell.neighbors,
  seed = 1234,
  ...
)
}
\arguments{
\item{Yg}{a numeric vector containing RNA expression values.}

\item{Xt}{a numeric vector containing chromatin accessibility.}

\item{Yj}{a numeric vector containing TF expression.}

\item{alternative}{a character string specifying the alternative hypothesis.
This must be one of "two.sided", "greater", and "less".}

\item{metacell.celltype}{a character vector specifying cell types of the metacells.}

\item{metacell.neighbors}{an integer vector containing indices of metacell
neighbors}

\item{seed}{an integer}

\item{...}{further arguments to be passed.}
}
\value{
A m-by-4 or m-by-5 matrix containing p-values where m represents the
number of meta cells.
}
\description{
This function computes sampling-based p-values representing the influence
of deleting a group of metacells on the linear interaction model.
}
