% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep2.R
\name{filterSeuratObject2}
\alias{filterSeuratObject2}
\title{Filter a Seurat object}
\usage{
filterSeuratObject2(object, tripod.object)
}
\arguments{
\item{object}{a Seurat object.}

\item{tripod.object}{a list object returned by {\code{\link{getObjectsForModelFit2}}}}
}
\value{
This function returns a Seurat object.
}
\description{
This function filters a Seurat object and an output
from {\code{\link{getObjectsForModelFit2}}}
to ensure that
TF binding motifs are kept only when genes encoding corresponding TFs are
present in the RNA-seq data and that the same set of genes is included
in all objects that are necessary for model fitting.
}
