% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{fitModel}
\alias{fitModel}
\title{Perform model fitting for a given target gene}
\usage{
fitModel(
  xymats,
  model.name,
  match.by = NULL,
  log = NULL,
  cap.at.quantile = 0.02,
  delta = 0.2,
  min.Xt = 5,
  min.Yj = 5,
  min.level1 = 10,
  min.level2 = 10,
  metacell.celltype.col = NULL,
  verbose = FALSE
)
}
\arguments{
\item{xymats}{a list returned by {\code{\link{getXYMatrices}}}.}

\item{model.name}{a character string specifying the model. This must be
"marginal", "conditional", "interaction", "TRIPOD".}

\item{match.by}{a character string specifying the variable to be matched.
This is necessary for "TRIPOD" only.}

\item{log}{a logical indicating whether to log-transform the variables.}

\item{cap.at.quantile}{a quantile at which variables are capped to remove outliers.
That is, values are capped if they are below cap.at.quantile
or above (1 - cap.at.quantile).
Set this to 0 if capping is unnecessary.}

\item{delta}{a numeric.}

\item{min.Xt}{an integer to set the minimum number of nonzero Xt.}

\item{min.Yj}{an integer to set the minimum number of nonzero Yj.}

\item{min.level1}{an integer to set the minimum number of pairs.}

\item{min.level2}{an integer to set the minimum number of pairs.}

\item{metacell.celltype.col}{a character vector specifying colors of the metacells
based on the cell types. This is necessary only for generating color-coded
scatter plots of pairs of metacells using {\code{\link{plotGenePeakTFScatter}}}.}

\item{verbose}{a logical indicating whether to print progress in "TRIPOD".}
}
\value{
A list containing some of the following elements. Note that the sets of
elements differ depending on the model.name input argument.
\item{gene.name}{a character string.}
\item{Yg}{ a vector containig capped RNA expression of the gene across metacells.}
\item{Xt}{a matrix containig capped chromatin accessibility of the ATAC peaks.}
\item{Yj}{a matrix containing capped TF RNA expression.
Note that TFs that do no have motifs in the region
around TSS are excluded.}
\item{peakxmotif.g}{a gene-specific binary matrix containing indicators of
whether ATAC peaks contains motifs.
rows and columns represent peaks and motifs, respectively.}
\item{nonzero.peakxmotif.g}{a matrix with two columns.
The first and second columns contain row and column numbers
in the peakmotif.g matrix where the entry is non-zero.}
\item{alpha.m}{a vector containing estimates of coefficient from the marginal model, Yg ~ Xt.}
\item{pval.alpha.m}{p-values associated with alpha.m.}
\item{beta.m}{a vector containing estimates of coefficient from the marginal model, Yg ~ Yj.}
\item{pval.beta.m}{p-values associated with beta.m.}
\item{gamma.m}{a matrix containing estimates of coefficient from the model, Yg ~ Xt:Yj.}
\item{pval.gamma.m}{p-values associated with gamma.m.}
\item{eta.m}{a matrix containing estimates of coefficient from the martinal model, Yj ~ Xt.}
\item{pval.eta.m}{p-values associated with eta.m.}
\item{cor.alpha}{Pearson correlation coefficient between Yg and Xt.}
\item{cor.beta}{Pearson correlation coefficient between Yg and Yj.}
\item{cor.gamma}{Pearson correlation coefficient between Yg and
the product of Xt and Yj.}
\item{cor.eta}{Pearson correlation coefficient between Yj and Xt.}
\item{alpha.c}{a matrix containing estimates of coefficient from the conditional model, Yg ~ Xt + Yj.}
\item{pval.alpha.c}{p-values associated with alpha.c.}
\item{beta.c}{a matrix containing estimates of coefficient from the conditional model, Yg ~ Xt + Yj.}
\item{pval.beta.c}{p-values associated with beta.c.}
\item{delta}{This applies to "conditional" and "TRIPOD" only.}
\item{alpha.i}{a matrix containing estimates of coefficient from the interation model.}
\item{pval.alpha.i}{p-values associated with alpha.i.}
\item{beta.i}{a matrix containing estimates of coefficient from the interation model.}
\item{pval.beta.i}{p-values associated with beta.i.}
\item{gamma.i}{a matrix containing estimates of coefficient from the interation model.}
\item{pval.gamma.i}{p-values associated with gamma.i.}
\item{coef.level1}{a matrix containing estimates of coefficient from TRIPOD level 1 test.}
\item{pval.level1}{p-values associated with coef.level1.}
\item{coef.level2}{a matrix containing estimates of coefficient from TRIPOD level 2 test.}
\item{pval.level2}{p-values associated with coef.level2.}
\item{match.by}{the matched variable. This applieds to "TRIPOD" only.}
}
\description{
fitModel() takes a list returned by {\code{\link{getXYMatrices}}}
as input, performs model fitting for a given target gene,
and returns a list object containing matrices of coefficients and p-values.
The (t, j) elements of the matrices
represent values for the t-th ATAC peak and the j-th TF.
}
