% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core2.R
\name{fitModel2}
\alias{fitModel2}
\title{Perform model fitting for a given target gene}
\usage{
fitModel2(
  xymats,
  model.name,
  match.by = NULL,
  log = NULL,
  cap.at.quantile = 0.02,
  delta = 0.2,
  min.Xp = 5,
  min.Yt = 5,
  min.level1 = 10,
  min.level2 = 10,
  metacell.celltype.col = NULL,
  verbose = FALSE
)
}
\arguments{
\item{xymats}{a list returned by {\code{\link{getXYMatrices}}}.}

\item{model.name}{a character string specifying the model. This must be
"marginal", "conditional", "interaction", "TRIPOD".}

\item{match.by}{a character string specifying the variable to be matched.
This is necessary for "TRIPOD" only.}

\item{log}{a logical indicating whether to log-transform the variables.}

\item{cap.at.quantile}{a quantile at which variables are capped to remove outliers.
That is, values are capped if they are below \code{cap.at.quantile}
or above (1 - \code{cap.at.quantile}).
Set this to 0 if capping is unnecessary.}

\item{delta}{a numeric.}

\item{min.Xp}{an integer to set the minimum number of metacells
with nonzero chromatin accessibility, \code{Xp}.}

\item{min.Yt}{an integer to set the minimum number of metacells
with nonzero TF expression, \code{Yt}.}

\item{min.level1}{an integer to set the minimum number of pairs.}

\item{min.level2}{an integer to set the minimum number of pairs.}

\item{metacell.celltype.col}{a character vector specifying colors of the metacells
based on the cell types. This is necessary only for generating color-coded
scatter plots of pairs of metacells using {\code{\link{plotGenePeakTFScatter}}}.}

\item{verbose}{a logical indicating whether to print progress in "TRIPOD".}
}
\value{
This function returns a list containing some of the following elements. Note that the sets of
elements differ depending on the \code{model.name} input argument.
\item{gene.name}{a character string.}
\item{ext.upstream}{an integer representing the window size in bp.}
\item{ext.downstream}{an integer representing the window size in bp.}
\item{Yg}{a vector containig capped RNA expression of the gene across metacells.}
\item{Xp}{a matrix containig capped chromatin accessibility of the ATAC peaks.}
\item{Yt}{a matrix containing capped TF RNA expression.
Note that TFs that do no have motifs in the region around TSS are excluded.}
\item{peak.motif.TF.g}{a gene-specific data frame containing mapping between
ATAC peaks, motif, and TFs. Each row contains a given ATAC peak region, a motif
that is present in the region, and the corresponding TF.}
\item{coef}{a list or data fram object containing coefficients and p-values.}
\item{delta}{This applies to "conditional" and "TRIPOD" only.}
\item{match.by}{the matched variable. This applieds to "TRIPOD" only.}
}
\description{
This function takes a list returned by {\code{\link{getXYMatrices}}}
as input, performs model fitting for a given target gene,
and returns a list or data fram object containing coefficients and p-values.
}
