% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{getCellTypeForMetacell}
\alias{getCellTypeForMetacell}
\title{Get cell types for metacells}
\usage{
getCellTypeForMetacell(
  object,
  celltype.col.name = "celltype",
  cluster.col.name = "seurat_clusters"
)
}
\arguments{
\item{object}{a Seurat object.}

\item{celltype.col.name}{a character string specifying the name of the
meta data column in the Seurat object containing cell types.}

\item{cluster.col.name}{a character string specifying the name of the
meta data column in the Seurat object containing cluster numbers.}
}
\value{
a character vector
}
\description{
Get cell types for metacells
}
