% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{getClusters}
\alias{getClusters}
\title{Get metacell assignment}
\usage{
getClusters(object, resolution, graph.name, ...)
}
\arguments{
\item{object}{a Seurat object.}

\item{resolution}{a numeric variable specifying the resolution for clustering.}

\item{graph.name}{a character string specifying the graph name. This must be
one of the graphs stored in the Seurat object.}

\item{...}{further arguments passed to {\code{\link{FindClusters}}}.}
}
\value{
This function returns a Seurat object.
}
\description{
This function takes an Seurat object as an input and
returns the object with a meta data column, \code{seurat_clusters}.
containing metacell assignment.
}
