% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{getColors}
\alias{getColors}
\title{Get color coding based on cell types}
\usage{
getColors(
  object,
  reduction = "umap.rna",
  celltype.col.name = "celltype",
  cluster.col.name = "seurat_clusters"
)
}
\arguments{
\item{object}{a Seurat object.}

\item{reduction}{a character string specifying a dimension reduction method.}

\item{celltype.col.name}{a character string specifying the name of the
meta data column in the Seurat object containing cell types.}

\item{cluster.col.name}{a character string specifying the name of the
meta data column in the Seurat object containing cluster numbers.}
}
\value{
a list of data frames
}
\description{
This function takes a Seurat object as input and
returns data frames containing mappings of colors to
single cells, metacells, and cell types.
}
