% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{getColorsForCellType}
\alias{getColorsForCellType}
\title{Get a mapping between cell types and colors}
\usage{
getColorsForCellType(ordered.celltypes, metacell.color.map)
}
\arguments{
\item{ordered.celltypes}{a character string.}

\item{metacell.color.map}{a data frame.}
}
\value{
a data frame
}
\description{
Get a mapping between cell types and colors
}
