% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{getColorsForMetacells}
\alias{getColorsForMetacells}
\title{Get a mapping between cell types and colors for meta cells}
\usage{
getColorsForMetacells(metacell.celltype, sc.color.map)
}
\arguments{
\item{metacell.celltype}{a character string}

\item{sc.color.map}{a data frame}
}
\value{
a data frame
}
\description{
Get a mapping between cell types and colors for meta cells
}
