% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{getColorsForSingleCell}
\alias{getColorsForSingleCell}
\title{Get a mapping between cell types and colors for single cells}
\usage{
getColorsForSingleCell(object, reduction, celltype.col.name)
}
\arguments{
\item{object}{a Seurat object.}

\item{reduction}{a character string specifying a dimension reduction method.}

\item{celltype.col.name}{a character string specifying the name of the meta
data column in the Seurat object containing cell types.}
}
\value{
a data frame
}
\description{
Get a mapping between cell types and colors for single cells
}
