% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{getDendrogram}
\alias{getDendrogram}
\title{Get a dendrogram}
\usage{
getDendrogram(metacell.matrix, metacell.celltype)
}
\arguments{
\item{metacell.matrix}{a matrix containing RNA expression values for genes
or chromatin accessibility values for ATAC peaks.
Rows and columns represent cell clusters and feature names, respectively.}

\item{metacell.celltype}{a character vector specifying cell types of the metacells.}
}
\value{
A dendrogram.
}
\description{
Get a dendrogram
}
