% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{getMetacellMatrices}
\alias{getMetacellMatrices}
\title{Get metacell matrices}
\usage{
getMetacellMatrices(
  object,
  cluster.name = "seurat_clusters",
  slot.rna = "counts",
  slot.atac = "counts",
  min.num = 0
)
}
\arguments{
\item{object}{a Seurat object.}

\item{cluster.name}{a character string specifying the cell clusters on which
metacells are based. This must be one of the column names in the meta data
of the Seurat object.}

\item{slot.rna}{a character string specifying the slot for the RNA modality.
This must be one of "counts" and "data".}

\item{slot.atac}{a character string specifying the slot for the ATAC modality.
This must be one of "counts" and "data."}

\item{min.num}{an integer to represent the minimum number of cells. Metacells
containing single cells fewer than this threshold will be removed.}
}
\value{
This function returns a list object containing the following elements.
\item{rna}{a matrix.}
\item{atac}{a matrix.}
}
\description{
This is a wrapper function of {\code{\link{getMetacellMatrix}}}.
}
