% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{getMetacellMatrix}
\alias{getMetacellMatrix}
\title{Get a metacell matrix}
\usage{
getMetacellMatrix(object, cluster.name = "seurat_clusters", assay)
}
\arguments{
\item{object}{a Seurat object.}

\item{cluster.name}{a character string specifying the cell clusters on which
metacells are based. This must be one of the column names in the meta data
of the Seurat object.}

\item{assay}{a character string specifying the assay. This should be
"RNA" and "ATAC" for RNA expression and chromatin accessibility, respectively.}
}
\value{
a matrix
}
\description{
This function takes a Seurat object as an input and returns a matrix
containing normalized RNA expression or chromatin accessibility per
metacell.
}
