% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{getMetacellNeighbor}
\alias{getMetacellNeighbor}
\title{Get metacell neighbors}
\usage{
getMetacellNeighbor(object, reduction, cluster.name = "seurat_clusters", k = 4)
}
\arguments{
\item{object}{a Seurat object.}

\item{reduction}{a character string specifying the reduced dimension data}

\item{cluster.name}{a character string specifying the cell clusters on which
metacells are based. This must be one of the column names in the meta data
of the Seurat object.}

\item{k}{an integer value setting the number of metacell neighbors to include}
}
\value{
A matrix.
}
\description{
Get metacell neighbors
}
