% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{getObjectsForModelFit}
\alias{getObjectsForModelFit}
\title{Get objects required for model fitting}
\usage{
getObjectsForModelFit(object, chr, convert = FALSE)
}
\arguments{
\item{object}{a Seurat object.}

\item{chr}{a character string specifying chromosomes. This would be
\code{paste0("chr", 1:22)} for the autosomes in human.}

\item{convert}{a logical variable to indicate whether the gene names need
to be converted. The current
version only supports conversion from human to mouse.
This is necessary when a human motif dataset is used for analyzing mouse data.}
}
\value{
This function returns a list object with the following elements.
\describe{
\item{transcripts.gr}{a GRanges object containing chromosomal coordinates
of annotated protein containing genes.}
\item{peaks.gr}{a GRanges object containing chromosomal coordinates
of the ATAC peak regions in the multiomic data.}
\item{motifxTF}{a matrix containing mapping between TFs and their binding motifs.}
\item{peakxmotif}{a binary sparse matrix containing indicators
as to whether a given motif is present in a given ATAC peak region.}
}
}
\description{
This function takes a Seurat object and returns a list object containing
elements that are required for model fitting.
}
