% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{getPeakGenePairs}
\alias{getPeakGenePairs}
\title{Get peak-gene pairs with significantly nonzero coefficients across genes}
\usage{
getPeakGenePairs(xymats.list, fdr.thresh, sign, model.name)
}
\arguments{
\item{xymats.list}{a list of lists (xymats objects) returned
by {\code{\link{fitModel}}}.}

\item{fdr.thresh}{a numerical scalar values setting a FDR threshold.}

\item{sign}{a character string specifying the sign.
This should be either "positive" or "negative".}

\item{model.name}{a character string specifying the model.}
}
\value{
a data frame
}
\description{
This is a wrapper function for {\code{\link{getPeakGenePairsForSingleGene}}}.
}
