% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{getPeakGenePairsForSingleGene}
\alias{getPeakGenePairsForSingleGene}
\title{Get peak-gene pairs with significantly nonzero coefficients for a given target gene}
\usage{
getPeakGenePairsForSingleGene(xymats, coef.name, pval.name, sign, fdr.thresh)
}
\arguments{
\item{xymats}{a list returned by {\code{\link{fitModel}}}.}

\item{coef.name}{a character string specifying the name of the coefficient.
This must be one of the names of the elements in the xymats object.}

\item{pval.name}{a character string specifying the name of the element in
the xymats object that stores the P value.}

\item{sign}{a character string specifying the sign.
This should be either "positive" or "negative".}

\item{fdr.thresh}{a numerical scalar values setting a FDR threshold.}
}
\value{
a data frame
}
\description{
This function takes a list returned by {\code{\link{fitModel}}}
as input and generate a data frame containing peak-gene pairs with significantly
positive or negative coefficients for a given target gene.
}
