% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{getTrios}
\alias{getTrios}
\title{Get trios with significantly nonzero coefficients across genes}
\usage{
getTrios(xymats.list, genes = NULL, fdr.thresh, sign, model.name, level = NULL)
}
\arguments{
\item{xymats.list}{a list of lists (xymats objects) returned
by {\code{\link{fitModel}}}.}

\item{genes}{a character vector containing the symbols of target genes
to be analyzed.}

\item{fdr.thresh}{a numerical scalar values setting a FDR threshold.}

\item{sign}{a character string specifying the sign.
This should be either "positive" or "negative".}

\item{model.name}{a character string specifying the model.}

\item{level}{an integer specifying whether level 1 or 2 test is used.}
}
\value{
a data frame
}
\description{
This is a wrapper function for {\code{\link{getTriosForSingleGene}}}.
}
