% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{getXYMatrices}
\alias{getXYMatrices}
\title{Create input matrices for model fitting for a given target gene}
\usage{
getXYMatrices(
  gene.name,
  ext.upstream,
  ext.downstream = NULL,
  transcripts.gr,
  peaks.gr,
  metacell.rna,
  metacell.peak,
  peakxmotif,
  motifxTF,
  metacell.celltype = NULL,
  metacell.celltype.col = NULL
)
}
\arguments{
\item{gene.name}{a character string containing the gene name. This must exist
in the "gene_name" column of the transcripts.gr object.}

\item{ext.upstream}{an integer representing the window size in bp.}

\item{ext.downstream}{an integer representing the window size in bp.}

\item{transcripts.gr}{a GRanges object containing annotated protein-coding transcripts.}

\item{peaks.gr}{a GRanges object containing ATAC peaks.}

\item{metacell.rna}{a matrix containing RNA expression values for genes.
Rows and columns represent cell clusters and gene names, respectively.}

\item{metacell.peak}{a matrix containing chromatin accessibility values for ATAC peaks.
Rows and columns represent cell clusters and ATAC peak positions, respectively.}

\item{peakxmotif}{a binary matrix containing indicators as to whether ATAC peaks contain motifs.}

\item{motifxTF}{a matrix containing names of motifs and TFs.}

\item{metacell.celltype}{a character vector specifying cell types of the metacells.}

\item{metacell.celltype.col}{a character vector specifying colors of the metacells
based on the cell types.}
}
\value{
A list with elements:
\item{gene.name}{a character string.}
\item{Yg}{ a vector containig RNA expression of the gene across metacells.}
\item{Xt}{a matrix containig chromatin accessibility of the ATAC peaks.}
\item{Yj}{a matrix containing TF RNA expression.
note that this is gene-specific since TFs that do no have motifs in the region
around TSS are excluded.}
\item{peakxmotif.g}{a gene-specific binary matrix containing indicators of
whether ATAC peaks contains motifs.
rows and columns represent peaks and motifs, respectively.}
\item{nonzero.peakxmotif.g}{a matrix with two columns.
The first and second columns contain row and column numbers
in the peakmotif.g matrix where the entry is non-zero.}
}
\description{
getXYMatrices() creates input matrices, Yg, X, Yj, for model fitting
for a given target gene. Unlike {\code{\link{getXYMatrices}}},
the cases where Yg and Yj are the same are excluded.
}
