% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core2.R
\name{getXYMatrices2}
\alias{getXYMatrices2}
\title{Create input matrices for model fitting for a given target gene}
\usage{
getXYMatrices2(
  gene.name,
  ext.upstream,
  ext.downstream = NULL,
  tripod,
  metacell
)
}
\arguments{
\item{gene.name}{a character string containing the gene name. This must exist
in the \code{gene_name} column of the \code{transcripts.gr} object.}

\item{ext.upstream}{an integer representing the window size in bp.}

\item{ext.downstream}{an integer representing the window size in bp.}

\item{tripod}{an object returned by {\code{\link{getObjectsForModelFit2}}}.}

\item{metacell}{an object returned by {\code{\link{getMetacellMatrices}}}.}
}
\value{
A list with elements:
\item{gene.name}{a character string.}
\item{Yg}{a vector containig RNA expression of the gene across metacells.}
\item{Xp}{a matrix containig chromatin accessibility of the ATAC peaks.}
\item{Yt}{a matrix containing TF RNA expression.
note that this is gene-specific since TFs that do no have motifs in the region
around TSS are excluded.}
\item{peak.motif.TF.g}{a gene-specific data frame containing mapping between
ATAC peaks, motif, and TFs. Each row contains a given ATAC peak region, a motif
that is present in the region, and the corresponding TF.}
}
\description{
This function creates input matrices, Yg, Xp, Yt, for model fitting
for a given target gene.
It allows for cases where a given TF corresponds to multiple motifs and
those where a given motif corresponds to multiple TFs.
Cases where Yg and Yt are the same are excluded.
}
