% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{label.savercat}
\alias{label.savercat}
\title{Cell-type labels of single cells obtained by SAVERCAT}
\format{
a character vector with 4881 elements
}
\usage{
label.savercat
}
\description{
Cell-type labels of single-cells multiome data of mouse embryonic brain
at day 18 obtained using SAVERCAT.
}
\keyword{datasets}
