% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{label.seurat}
\alias{label.seurat}
\title{Cell-type labels of single cells obtained by Seurat}
\format{
a character vector with 4881 elements
}
\usage{
label.seurat
}
\description{
Cell-type labels of single-cells multiome data of mouse embryonic brain
at day 18 obtained using Seurat.
}
\keyword{datasets}
