% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metacell.celltype.col}
\alias{metacell.celltype.col}
\title{Colors assigned to metacells}
\format{
a character vector with 84 elements
}
\usage{
metacell.celltype.col
}
\description{
Cell type colors of metacells identified in single-cell multiome data
of mouse embryonic brain at day 18.
}
\keyword{datasets}
