% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metacell.celltype}
\alias{metacell.celltype}
\title{Cell types of metacells}
\format{
a character vector with 84 elements
}
\usage{
metacell.celltype
}
\description{
Assigned cell types of metacells identified in single-cell multiome data
of mouse embryonic brain at day 18.
}
\keyword{datasets}
