% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metacell.peak}
\alias{metacell.peak}
\title{Chromatin accessibility in metacells}
\format{
a matrix object with 84 rows and 139083 columns:
\describe{
\item{row}{metacell indices}
\item{column}{ATAC peak chromosomal coordinates}
}
}
\usage{
metacell.peak
}
\description{
Normalized Chromatin accessibility of 139083 ATAC peaks in 84 metacells.
identified in single-cell multiome data of mouse embryonic brain at day 18.
}
\keyword{datasets}
