% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metacell.rna}
\alias{metacell.rna}
\title{RNA expression in metacells}
\format{
a matrix with 84 rows and 14476 columns:
\describe{
\item{row}{metacell indices}
\item{column}{gene symbols}
}
}
\usage{
metacell.rna
}
\description{
Normalized RNA expression levels of 14476 genes in 84 metacells identified
in single-cell multiome data of mouse embryonic brain at day 18.
}
\keyword{datasets}
