% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motifxTF}
\alias{motifxTF}
\title{TFs and their binding motifs}
\format{
a matrix object with 351 rows and 2 columns:
\describe{
\item{motif}{JASPAR motif IDs}
\item{TF}{gene types}
}
}
\usage{
motifxTF
}
\description{
TF binding motif data for 351 TFs based on the JASPAR2020 database.
}
\keyword{datasets}
