% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{optimizeResolution}
\alias{optimizeResolution}
\title{Optimize cluster resolutions}
\usage{
optimizeResolution(
  object,
  assay.name,
  graph.name = NULL,
  algorithm = NULL,
  resolutions = seq(10, 35, 5),
  min.num,
  ...
)
}
\arguments{
\item{object}{a Seurat object.}

\item{assay.name}{a character string specifying the assay name. This must
be one of "SCT", "ATAC", or "WNN".}

\item{graph.name}{a character string specifying the graph name. This must be
one of the graphs stored in the Seurat object.}

\item{algorithm}{an integer representing a clustering algorithm. See
{\code{\link{FindClusters}}} in the Seurat package.}

\item{resolutions}{a numerical vector specifying resolutions to be examined.}

\item{min.num}{an integer setting a threshold for the minimum number of
cells per cluster.}

\item{...}{further arguments to be passed to {\code{\link{FindClusters}}}.}
}
\value{
a data frame with three columns:
\describe{
\item{resolusion}{the resolution examined}
\item{num_clusters}{the number of clusters}
\item{num_below}{the number of clusters with fewer than
the threshold number of single cells}
}
}
\description{
This function takes a Seurat object as input and returns a data frame
containing cluster numbers at specified resolutions.
}
