% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc.metacell.celltype.col}
\alias{pbmc.metacell.celltype.col}
\title{Colors assigned to metacells in PBMC}
\format{
a character vector with 80 elements
}
\usage{
pbmc.metacell.celltype.col
}
\description{
Cell type colors of metacells identified in single-cell multiome data
of human peripheral blood mononuclear cells (PBMC).
}
\keyword{datasets}
