% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc.metacell.celltype}
\alias{pbmc.metacell.celltype}
\title{Cell types of metacells in PBMC}
\format{
a character vector with 80 elements
}
\usage{
pbmc.metacell.celltype
}
\description{
Assigned cell types of metacells identified in single-cell multiome data
of human peripheral blood mononuclear cells (PBMC).
}
\keyword{datasets}
