% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc.metacell.peak}
\alias{pbmc.metacell.peak}
\title{Chromatin accessibility in metacells in PBMC}
\format{
a matrix object with 80 rows and 106056 columns:
\describe{
\item{row}{metacell indices}
\item{column}{ATAC peak chromosomal coordinates}
}
}
\usage{
pbmc.metacell.peak
}
\description{
Normalized Chromatin accessibility of 106056 ATAC peaks in 80 metacells.
identified in single-cell multiome data of human peripheral blood mononuclear cells (PBMC).
}
\keyword{datasets}
