% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc.metacell.rna}
\alias{pbmc.metacell.rna}
\title{RNA expression in metacells in PBMC}
\format{
a matrix with 80 rows and 14508 columns:
\describe{
\item{row}{metacell indices}
\item{column}{gene symbols}
}
}
\usage{
pbmc.metacell.rna
}
\description{
Normalized RNA expression levels of 14508 genes in 80 metacells identified
in single-cell multiome data of human peripheral blood mononuclear cells (PBMC).
}
\keyword{datasets}
