% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc.motifxTF}
\alias{pbmc.motifxTF}
\title{TFs and their binding motifs for analyzing the PBMC data}
\format{
a matrix object with 342 rows and 2 columns:
\describe{
\item{motif}{JASPAR motif IDs}
\item{TF}{gene types}
}
}
\usage{
pbmc.motifxTF
}
\description{
TF binding motif data for 342 TFs based on the JASPAR2020 database.
}
\keyword{datasets}
