% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc.peaks.gr}
\alias{pbmc.peaks.gr}
\title{ATAC peaks in PBMC}
\format{
a GRanges object with 106056 elements
}
\source{
\url{https://support.10xgenomics.com/single-cell-multiome-atac-gex/datasets}
}
\usage{
pbmc.peaks.gr
}
\description{
A GRanges object containing 106056 ATAC peaks identified
in single-cell multiome data of human peripheral blood mononuclear cells (PBMC).
}
\keyword{datasets}
