% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc.peakxmotif}
\alias{pbmc.peakxmotif}
\title{The presence and absence of TF binding motifs in ATAC peaks in PBMC}
\format{
a dgCMatrix matrix with 106056 rows and 342 columns:
\describe{
\item{row}{ATAC peak chromosomal coordinates}
\item{column}{JASPAR motif IDs}
}
}
\usage{
pbmc.peakxmotif
}
\description{
A binary sparse matrix indicating whether TF binding motifs are present
in ATAC peaks identified in single-cell multiome data of
human peripheral blood mononuclear cells (PBMC).
}
\keyword{datasets}
