% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc.transcripts.gr}
\alias{pbmc.transcripts.gr}
\title{Transcripts in PBMC}
\format{
a GRanges object with 14508 elements:
\describe{
\item{gene_name}{gene symbols}
\item{gene_biotype}{gene types}
}
}
\source{
\url{https://support.10xgenomics.com/single-cell-multiome-atac-gex/datasets}
}
\usage{
pbmc.transcripts.gr
}
\description{
A GRanges object containing 14508 protein_coding genes detected
in single-cell multiome data of human peripheral blood mononuclear cells (PBMC).
}
\keyword{datasets}
