% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peaks.gr}
\alias{peaks.gr}
\title{ATAC peaks}
\format{
a GRanges object with 139083 elements
}
\source{
\url{https://support.10xgenomics.com/single-cell-multiome-atac-gex/datasets}
}
\usage{
peaks.gr
}
\description{
A GRanges object containing 139083 ATAC peaks identified
in single-cell multiome data of mouse embryonic brain at day 18.
}
\keyword{datasets}
