% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peakxmotif}
\alias{peakxmotif}
\title{The presence and absence of TF binding motifs in ATAC peaks}
\format{
a dgCMatrix matrix with 139083 rows and 351 columns:
\describe{
\item{row}{ATAC peak chromosomal coordinates}
\item{column}{JASPAR motif IDs}
}
}
\usage{
peakxmotif
}
\description{
A binary sparse matrix indicating whether TF binding motifs are present
in ATAC peaks identified in single-cell multiome data of mouse embryonic brain
at day 18.
}
\keyword{datasets}
