% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{performBHMatrix}
\alias{performBHMatrix}
\title{Perform the Benjamini and Hochberg FDR procedure on a matrix}
\usage{
performBHMatrix(pvals, fdr.thresh = 0.01, do.plot = NULL)
}
\arguments{
\item{pvals}{a matrix containing p-values.}

\item{fdr.thresh}{a numerical scalar specifying the FDR threshold.}

\item{do.plot}{a logical scalar indicating whether to
produce a diagnostic plot.}
}
\value{
a logical matrix indicating whether the null hypothesis
has been rejected for each element.
}
\description{
Perform the Benjamini and Hochberg FDR procedure on a matrix
}
