% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{performBHVector}
\alias{performBHVector}
\title{Perform the Benjamini and Hochberg FDR procedure on a vector}
\usage{
performBHVector(pval.c, fdr.thresh = 0.01)
}
\arguments{
\item{pval.c}{a numerical vector containing p-values.}

\item{fdr.thresh}{a numerical scalar specifying the FDR threshold.}
}
\value{
a list of two elements:
\describe{
\item{pval.c.adj}{a numerical vector containing adjusted p-values}
\item{which.reject}{a logical vector indicating whether the null hypothesis
has been rejected for each element.}
}
}
\description{
Perform the Benjamini and Hochberg FDR procedure on a vector
}
