% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{performMatchedTest}
\alias{performMatchedTest}
\title{Perform matched test}
\usage{
performMatchedTest(
  Yg,
  Xt,
  Yj,
  Z = NULL,
  match.by,
  cap.at.quantile,
  delta,
  min.level1,
  min.level2,
  metacell.celltype.col = NULL,
  verbose = FALSE
)
}
\arguments{
\item{Yg}{a numeric vector containing RNA expression values
for a given gene g.}

\item{Xt}{a numeric vector containing chromatin accessibility
for a given peak t.}

\item{Yj}{a numeric vector containing TF expression for a given TF j.}

\item{Z}{a matrix containing covariates.}

\item{match.by}{a character string specifying the variable to be matched.
This is necessary for "TRIPOD" only.}

\item{cap.at.quantile}{a quantile at which variables are capped to remove outliers.
That is, values are capped if they are below cap.at.quantile
or above (1 - cap.at.quantile).
Set this to 0 if capping is unnecessary.}

\item{delta}{a numeric.}

\item{min.level1}{an integer to set the minimum number of pairs.}

\item{min.level2}{an integer to set the minimum number of pairs.}

\item{metacell.celltype.col}{a character vector specifying colors of the metacells
based on the cell types. This is necessary only for generating color-coded
scatter plots of pairs of metacells using {\code{\link{plotGenePeakTFScatter}}}.}

\item{verbose}{a logical indicating whether to print progress in "TRIPOD".}
}
\value{
a list
}
\description{
This function takes gene expression and chromatin accessibility vectors for
a given trio and performs matching-based nonparametric tests.
}
