% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core2.R
\name{performMatchedTest2}
\alias{performMatchedTest2}
\title{Perform matched test}
\usage{
performMatchedTest2(
  Yg,
  Xp,
  Yt,
  Z = NULL,
  match.by,
  cap.at.quantile,
  delta,
  min.level1,
  min.level2,
  metacell.celltype.col = NULL,
  verbose = FALSE
)
}
\arguments{
\item{Yg}{a numeric vector containing RNA expression values
for a given gene g.}

\item{Xp}{a numeric vector containing chromatin accessibility
for a given peak t.}

\item{Yt}{a numeric vector containing TF expression for a given TF j.}

\item{Z}{a matrix containing covariates.}

\item{match.by}{a character string specifying the variable to be matched.
This is necessary for "TRIPOD" only.}

\item{cap.at.quantile}{a quantile at which variables are capped to remove outliers.
That is, values are capped if they are below \code{cap.at.quantile}
or above (1 - \code{cap.at.quantile}).
Set this to 0 if capping is unnecessary.}

\item{delta}{a numeric.}

\item{min.level1}{an integer to set the minimum number of pairs.}

\item{min.level2}{an integer to set the minimum number of pairs.}

\item{metacell.celltype.col}{a character vector specifying colors of the metacells
based on the cell types. This is necessary only for generating color-coded
scatter plots of pairs of metacells using {\code{\link{plotGenePeakTFScatter}}}.}

\item{verbose}{a logical indicating whether to print progress in "TRIPOD".}
}
\value{
A list containing the following elements.
\item{matched.cells}{a data frame with columns containing metacell indices
in matche pairs, differences in the outcome Yg,
differences in the varied predictor V,
and the average of the matched predictor U}
\item{match.by}{a character string specifying the matched variable.}
\item{coef.level1}{a numerical scalar representing an estimate of coefficient
from TRIPOD level 1 test.}
\item{pval.level1}{a numerical scalar representing a p-value associated with
coef.level1}
\item{coef.level2}{a numerical scalar representing an estimate of coefficient
from TRIPOD level 2 test.}
\item{pval.level2}{a numerical scalar representing a p-value associated with
coef.level2}
\item{matched.col1}{a character vector specifying colors of the first metacells
in the pairs based on the cell types.}
\item{matched.col2}{a character vector specifying colors of the second metacells
in the pairs based on the cell types.}
}
\description{
This function takes gene expression and chromatin accessibility vectors for
a given trio and performs matching-based nonparametric tests.
}
