% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{performRNAPrediction}
\alias{performRNAPrediction}
\title{Perform RNA prediction for a given target gene}
\usage{
performRNAPrediction(
  xymats,
  model.name,
  log = NULL,
  cap.at.quantile = 0.02,
  seed = NULL
)
}
\arguments{
\item{xymats}{a list returned by {\code{\link{getXYMatrices}}}.}

\item{model.name}{a character string specifying the model. This must be
one of "gene.activity", "peak.LASSO", or "peak.TF.LASSO".}

\item{log}{a logical indicating whether to log-transform the variables.
This must be one of "Yg", "Xt", and "Yj".}

\item{cap.at.quantile}{a quantile at which variables are capped to remove outliers.
Set this to 0 if capping is unnecessary.}

\item{seed}{an integer.}
}
\value{
a list
}
\description{
This function takes a list returned by {\code{\link{getXYMatrices}}}
as input and performs RNA prediction for a given target gene.
}
