% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCooksD}
\alias{plotCooksD}
\title{Plot Cook's distance}
\usage{
plotCooksD(d.list, metacell.celltype.col, ...)
}
\arguments{
\item{d.list}{a list returned by {\code{\link{getCooksD}}}}

\item{metacell.celltype.col}{a character vector specifying colors of the metacells
based on the cell types.}

\item{...}{further arguments to be passed.}
}
\value{
A ggplot object.
}
\description{
Plot Cook's distance
}
