% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDFFIT}
\alias{plotDFFIT}
\title{Plot DFFIT}
\usage{
plotDFFIT(d.list, metacell.celltype.col, ...)
}
\arguments{
\item{d.list}{a list returned by {\code{\link{getDFFIT}}}}

\item{metacell.celltype.col}{a character vector specifying colors of the metacells
based on the cell types.}

\item{...}{further arguments to be passed.}
}
\value{
A ggplot object.
}
\description{
Plot DFFIT
}
