% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotFeatureForNeighbor}
\alias{plotFeatureForNeighbor}
\title{Generate a UMAP plot colored by sampling-based P values}
\usage{
plotFeatureForNeighbor(delta.coeff.pval, object, reduction, coefficient, ...)
}
\arguments{
\item{delta.coeff.pval}{a data frame returned by {\code{\link{deleteNeighbor}}}}

\item{object}{a Seurat object.}

\item{reduction}{a character string specifying the dimension reduction method.
The coordinates should be stored in the Seurat object.}

\item{coefficient}{a character string specifying the coefficient to be plotted.
This must be one of "Xt", "Yj", and "Xt:Yj".
If {\code{\link{deleteNeighbor}}} was run with an argument "two.sided",
this can be set to fitted values, "Yg".}

\item{...}{further arguments to be passed.}
}
\value{
A ggplot object.
}
\description{
This function generates a UMAP plot colored by sampling-based P values
for deleting a group of metacells.
}
