% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGenePeakTFDot}
\alias{plotGenePeakTFDot}
\title{Generate dot plots for a trio}
\usage{
plotGenePeakTFDot(
  object,
  gene.name,
  peak.name,
  TF.name,
  group.by,
  size.title = 15,
  size.axis = 15,
  size.legend = 15,
  ncol = 3,
  ...
)
}
\arguments{
\item{object}{a Seurat object.}

\item{gene.name}{a character string representing a target gene name.}

\item{peak.name}{a character string containing a peak name,
such as "chr15-79201373-79203205".}

\item{TF.name}{a character string containing a TF name.}

\item{group.by}{a character string specifying a meta data column
in the Seurat object containing groups on which plot stratification is based.}

\item{size.title}{a numerical scalar specifying the size of the titles
of the plots.}

\item{size.axis}{a numerical scalar specifying the size of the axis titles.}

\item{size.legend}{a numerical scalar specifying the size of the legend titles.}

\item{ncol}{an integer specifying the number of columns for arranging plots.}

\item{...}{further arguments passed to {\code{\link{DotPlot}}}.}
}
\description{
This is a wrapper function of {\code{\link{DotPlot}}} from the Seurat
package for generating dot plots of target gene expression,
chromatin accessibility of an ATAC peak region, and TF expression
across specificed groups of cells.
}
