% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGenePeakTFScatter}
\alias{plotGenePeakTFScatter}
\title{Generate scatter plots for analyzing regulatory relationships}
\usage{
plotGenePeakTFScatter(
  xymats,
  peak.name,
  TF.name,
  to.plot,
  level = NULL,
  match.by = NULL,
  cap.at.quantile = 0.02,
  delta = 0.2,
  min.Xt = 5,
  min.Yj = 5,
  min.level1 = 10,
  min.level2 = 10,
  metacell.celltype = NULL,
  metacell.celltype.col = "black"
)
}
\arguments{
\item{xymats}{a list returned by {\code{\link{getXYMatrices}}}.}

\item{peak.name}{a character string containing a peak name,
such as "chr15-79201373-79203205".}

\item{TF.name}{a character string containing a TF name.}

\item{to.plot}{a character string specifying the model to be analyzed.
This must be one of "marginal.Yg.Xt", "marginal.Yg.Yj", "marginal.Yj.Xt",
"product", "conditional.on.Yj", "conditional.on.Xt", "interaction", or "TRIPOD".}

\item{level}{a numerical value specifying the level for TRIPOD. This must be
either 1 or 2.}

\item{match.by}{a character string specifying the variable to be matched.
This is necessary for "TRIPOD" only.}

\item{cap.at.quantile}{a quantile at which variables are capped to remove outliers.
That is, values are capped if they are below cap.at.quantile
or above (1 - cap.at.quantile).
Set this to 0 if capping is unnecessary.}

\item{delta}{a numeric.}

\item{min.Xt}{an integer to set the minimum number of nonzero Xt.}

\item{min.Yj}{an integer to set the minimum number of nonzero Yj.}

\item{min.level1}{an integer to set the minimum number of pairs.}

\item{min.level2}{an integer to set the minimum number of pairs.}

\item{metacell.celltype}{a character vector specifying cell types of the metacells.}

\item{metacell.celltype.col}{a character vector representing cell types of metacells.}
}
\description{
This function generated scatter plots of Yg, Xt, Yj, difference, or partial residuals.
Pearson correlation coefficient and associated P value are also printed.
}
