% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGenePeakTFUMAP}
\alias{plotGenePeakTFUMAP}
\title{Generate UMAP plots for a trio}
\usage{
plotGenePeakTFUMAP(
  object,
  gene.name,
  peak.name,
  TF.name,
  reduction = "wnn.umap",
  label = FALSE,
  size.title = 15,
  size.axis = 15,
  ncol = 3,
  ...
)
}
\arguments{
\item{object}{a Seurat object.}

\item{gene.name}{a character string representing a target gene name.}

\item{peak.name}{a character string containing a peak name,
such as "chr15-79201373-79203205".}

\item{TF.name}{a character string containing a TF name.}

\item{reduction}{a character string specifying the dimension reduction method.
The coordinates should be stored in the Seurat object.}

\item{label}{a logical scalar indicating whether to show cluster labels.}

\item{size.title}{a numerical scalar specifying the size of the titles
of the plots.}

\item{size.axis}{a numerical scalar specifying the size of the axis titles.}

\item{ncol}{an integer specifying the number of columns for arranging plots.}

\item{...}{further arguments passed to {\code{\link{FeaturePlot}}}.}
}
\description{
This is a wrapper function of {\code{\link{FeaturePlot}}} from the Seurat
package for generating UMAP plots, in which colors represent degrees of
target gene expression,
chromatin accessibility of an ATAC peak region, and TF expression
across cells.
}
