% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotInfluenceForCellType}
\alias{plotInfluenceForCellType}
\title{Plot sampling-based P-values for deleting a celltype}
\usage{
plotInfluenceForCellType(
  delta.coeff.pval,
  coefficient,
  threshold = 0.05,
  metacell.celltype,
  metacell.celltype.col,
  ordered.celltype
)
}
\arguments{
\item{delta.coeff.pval}{a data frame returned by {\code{\link{deleteNeighbor}}}}

\item{coefficient}{a character string specifying the coefficient to be plotted.
This must be one of "Xt", "Yj", and "Xt:Yj".
If {\code{\link{deleteNeighbor}}} was run with an argument "two.sided",
this can be set to fitted values, "Yg".}

\item{threshold}{a value of nominal p-value at which a horizontal line is drawn.}

\item{metacell.celltype}{a character vector specifying cell types of the metacells.}

\item{metacell.celltype.col}{a character vector specifying colors of the metacells
based on the cell types.}

\item{ordered.celltype}{a chracter vector containing cell types. This
specifies the order of the cell types in the plot.}
}
\value{
A ggplot object.
}
\description{
Plot sampling-based P-values for deleting a celltype
}
