% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotInfluenceForNeighbor}
\alias{plotInfluenceForNeighbor}
\title{Plot sampling-based P values for deleting a group of metacells}
\usage{
plotInfluenceForNeighbor(
  delta.coeff.pval,
  coefficient,
  metacell.celltype.col,
  threshold = 0.05,
  ...
)
}
\arguments{
\item{delta.coeff.pval}{a data frame returned by {\code{\link{deleteNeighbor}}}}

\item{coefficient}{a character string specifying the coefficient to be plotted.
This must be one of "Xt", "Yj", or "Xt:Yj".
If {\code{\link{deleteNeighbor}}} was run with an argument "two.sided",
this can be set to fitted values, "Yg".}

\item{metacell.celltype.col}{a character vector specifying colors of the metacells.
based on the cell types.}

\item{threshold}{a value of nominal P value at which a horizontal line is drawn}

\item{...}{further arguments to be passed.}
}
\value{
a ggplot object
}
\description{
Plot sampling-based P values for deleting a group of metacells
}
