% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotRNAPrediction}
\alias{plotRNAPrediction}
\title{Generate scatter plots for analyzing RNA prediction}
\usage{
plotRNAPrediction(
  xymats,
  metacell.celltype = NULL,
  metacell.celltype.col = "black",
  ...
)
}
\arguments{
\item{xymats}{a list object returned by {\code{\link{performRNAPrediction}}}.}

\item{metacell.celltype}{a character vector specifying cell types of the metacells.}

\item{metacell.celltype.col}{a character vector representing cell types of metacells.}

\item{...}{further arguments passed to {\code{\link{plot}}}}
}
\description{
This function takes an list object returned by
{\code{\link{performRNAPrediction}}} and generates scatter plots
comparing predicted and observed RNA expression values.
Pearson correlation coefficient is also printed.
}
