% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{processSeuratObject}
\alias{processSeuratObject}
\title{Process a Seurat object}
\usage{
processSeuratObject(object, dim.rna = 1:50, dim.atac = 2:50, verbose = TRUE)
}
\arguments{
\item{object}{a Seurat object.}

\item{dim.rna}{a integer vector passed to {\code{\link{RunUMAP}}},
{\code{\link{FindNeighbors}}}, and {\code{\link{FindMultiModalNeighbors}}}
for the RNA modality.}

\item{dim.atac}{a integer vector passed to {\code{\link{RunUMAP}}},
{\code{\link{FindNeighbors}}}, and {\code{\link{FindMultiModalNeighbors}}}
for the ATAC modality.}

\item{verbose}{a logical variable.}
}
\value{
This function returns a Seurat object.
}
\description{
This is a wrapper function to perform normalization, dimension reduction,
nearest-neighbor graph construction, and UMAP embedding for the RNA and
ATAC modalities. It also calculates a weighted nearest neighbor (WNN) graph.
}
