% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{removeSmallMetacells}
\alias{removeSmallMetacells}
\title{Remove small metacells}
\usage{
removeSmallMetacells(object, min.num)
}
\arguments{
\item{object}{a Seurat object.}

\item{min.num}{an integer to represent the minimum number of cells. Metacells
containing single cells fewer than this threshold will be removed.
This must be set to the same value as the input argument to
{\code{\link{getMetacellMatrices}}}.}
}
\value{
A Seurat object.
}
\description{
This functions removes cells corresponding to metacells with cells fewer than
a threshold from a Seurat object. {\code{\link{getMetacellMatrices}}} must be
run before running this function.
}
