% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{testInfluence}
\alias{testInfluence}
\title{Get sampling-based p-values}
\usage{
testInfluence(
  Yg,
  Xt,
  Yj,
  alternative,
  metacell.rm,
  metacell.celltype,
  plot.histogram = FALSE,
  nsamp = 10000,
  seed = 1234,
  ...
)
}
\arguments{
\item{Yg}{a numeric vector containing RNA expression values.}

\item{Xt}{a numeric vector containing chromatin accessibility.}

\item{Yj}{a numeric vector containing TF expression.}

\item{alternative}{a character string specifying the alternative hypothesis.
This must be one of "two.sided", "greater", and "less".}

\item{metacell.rm}{an integer vector containing indices of metacells
to be removed.}

\item{metacell.celltype}{a character vector specifying cell types of the metacells.}

\item{plot.histogram}{an indicator as to whether to generate histograms}

\item{nsamp}{an integer indicating the number of sampling}

\item{seed}{an integer}

\item{...}{further arguments passed to {\code{\link{hist}}}.}
}
\value{
A numerical vector containing p-values.
}
\description{
This function computes sampling-based p-values representing influences
of deleting metacells in a linear regression model.
}
