% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transcripts.gr}
\alias{transcripts.gr}
\title{Transcripts}
\format{
a GRanges object with 14476 elements:
\describe{
\item{gene_name}{gene symbols}
\item{gene_biotype}{gene types}
}
}
\source{
\url{https://support.10xgenomics.com/single-cell-multiome-atac-gex/datasets}
}
\usage{
transcripts.gr
}
\description{
A GRanges object containing 14476 protein_coding genes detected
in single-cell multiome data of mouse embryonic brain at day 18.
}
\keyword{datasets}
