% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transformCoef}
\alias{transformCoef}
\title{Transform coefficients to make them easier to plot}
\usage{
transformCoef(coefs)
}
\arguments{
\item{coefs}{a numerical matrix or vector.}
}
\value{
a numerical matrix or vector.
}
\description{
Transform coefficients to make them easier to plot
}
