% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transformPvalue}
\alias{transformPvalue}
\title{Transform p-values to make them easier to plot}
\usage{
transformPvalue(pvals)
}
\arguments{
\item{pvals}{a numerical matrix or vector.}
}
\value{
a numerical matrix or vector.
}
\description{
Transform p-values to make them easier to plot
}
