% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{getMetacellMatrix}
\alias{getMetacellMatrix}
\title{Get a metacell matrix}
\usage{
getMetacellMatrix(object, cluster.name, assay, slot)
}
\arguments{
\item{object}{a Seurat object.}

\item{cluster.name}{a character string specifying the cell clusters on which
metacells are based. This must be one of the column names in the meta data
of the Seurat object.}

\item{assay}{a character string specifying the assay. This must be one of
"RNA", "SCT", and "ATAC".}

\item{slot}{a character string specifying the slot. This must be one of
"counts" or "data".}
}
\value{
a matrix
}
\description{
This function takes a Seurat object as an input and returns a matrix
containing normalized RNA expression or chromatin accessibility per
metacell.
}
