% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{getTrios}
\alias{getTrios}
\title{Get trios with significantly nonzero coefficients across genes}
\usage{
getTrios(xymats.list, genes = NULL, fdr.thresh, sign, model.name, level = NULL)
}
\arguments{
\item{xymats.list}{a list of lists (xymats objects) returned
by {\code{\link{fitModel}}}.}

\item{genes}{a character vector containing the symbols of target genes
to be analyzed.}

\item{fdr.thresh}{a numerical scalar setting a FDR threshold.}

\item{sign}{a character string specifying the sign.
This should be either "positive" or "negative".}

\item{model.name}{a character string specifying the model.}

\item{level}{an integer specifying whether level 1 or 2 test is used.}
}
\value{
A data frame with 8 columns containing the following:
\item{gene}{a character string containing the name of the target gene.}
\item{peak_num}{an integer representing a peak index.
The mapping between the indices and the peak names are the same as that
in the xymats object.}
\item{TF_num}{an integer representing a TF index.
The mapping between the indices and the TF names are the same as that
in the xymats object.}
\item{peak}{a chracter string representingthe name of the peak.}
\item{TF}{a chracter string representing the name of the peak.}
\item{coef}{a numerical scalar representing the value of the coefficient}
\item{pval}{a numerical scalar representing the nominal p value}
\item{adj}{a numerical scalar representing the BH-adjusted p value}
}
\description{
This is a wrapper function for {\code{\link{getTriosForSingleGene}}}.
}
