% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{performRNAPrediction}
\alias{performRNAPrediction}
\title{Perform RNA prediction for a given target gene}
\usage{
performRNAPrediction(
  xymats,
  model.name,
  log = NULL,
  cap.at.quantile = 0.02,
  seed = NULL
)
}
\arguments{
\item{xymats}{a list returned by {\code{\link{getXYMatrices}}}.}

\item{model.name}{a character string specifying the model. This must be
one of "gene.activity", "peak.LASSO", or "peak.TF.LASSO".}

\item{log}{a logical indicating whether to log-transform the variables.
This must be one of "Yg", "Xt", and "Yj".}

\item{cap.at.quantile}{a quantile at which variables are capped to remove outliers.
Set this to 0 if capping is unnecessary.}

\item{seed}{an integer.}
}
\value{
A list containing some of the following elements.
\item{gene.name}{a character string.}
\item{Yg}{ a vector containig capped RNA expression of the gene across metacells.}
\item{Xt}{a matrix containig capped chromatin accessibility of the ATAC peaks.}
\item{Yj}{a matrix containing capped TF RNA expression.
Note that TFs that do no have motifs in the region
around TSS are excluded.}
\item{peakxmotif.g}{a gene-specific binary matrix containing indicators of
whether ATAC peaks contains motifs.
rows and columns represent peaks and motifs, respectively.}
\item{nonzero.peakxmotif.g}{a matrix with two columns.
The first and second columns contain row and column numbers
in the peakmotif.g matrix where the entry is non-zero.}
\item{Y.pred}{A numerical vector containing predicted values of Yg.}
\item{cor}{A numerical scalar representing Pearson correlation between
the predicted and observed Yg values}
}
\description{
This function takes a list returned by {\code{\link{getXYMatrices}}}
as input and performs RNA prediction for a given target gene.
}
