% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotTree}
\alias{plotTree}
\title{Plot sampling-based p-values for deleting a cell type}
\usage{
plotTree(
  dend,
  delta.coeff.pval,
  coefficient,
  metacell.celltype,
  metacell.celltype.col,
  threshold = 0.05,
  ...
)
}
\arguments{
\item{dend}{a dendrogram.}

\item{delta.coeff.pval}{a data frame returned by {\code{\link{deleteNeighbor}}}}

\item{coefficient}{a character string specifying the coefficient to be plotted.
This must be one of "Xt", "Yj", and "Xt:Yj".
If {\code{\link{deleteNeighbor}}} was run with an argument "two.sided",
this can be set to fitted values, "Yg".}

\item{metacell.celltype}{a character vector specifying cell types of the metacells.}

\item{metacell.celltype.col}{a character vector specifying colors of the metacells
based on the cell types.}

\item{threshold}{a value of nominal p-value at which a horizontal line is drawn.}

\item{...}{further arguments to be passed.}
}
\value{
A ggplot object.
}
\description{
Plot sampling-based p-values for deleting a cell type
}
